requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
cssselect==1.3.0
distro==1.9.0
exceptiongroup==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.74
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
lxml==6.0.1
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
openai==1.101.0
orjson==3.11.2
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
playwright==1.54.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2025.7.34
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.35.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
zstandard==0.24.0
